ai/requirements.txt
fastapi
fastapi[standard]
uvicorn
psycopg2-binary
pydantic
langchain
langchain-openai
openai
requests
python-dotenv
